v1/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
langgraph>=0.0.15
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.5.2
python-multipart>=0.0.6
sqlparse>=0.5.3
